% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{get_nhp_result_sets}
\alias{get_nhp_result_sets}
\title{List All NHP Results Data Files and their Metadata}
\usage{
get_nhp_result_sets(
  auth_token = azkit::get_auth_token(),
  table_ep = Sys.getenv("AZ_TABLE_EP"),
  runs_table_name = Sys.getenv("AZ_TABLE_NAME"),
  allowed_datasets = get_user_allowed_datasets(NULL)
)
}
\arguments{
\item{container_results}{Name of the blob_container/storage_container object
that stores results files.}
}
\value{
A data.frame. As many rows as there are files in \code{container}. As many
columns as there are metadata elements, plus the file path.
}
\description{
List All NHP Results Data Files and their Metadata
}
\details{
Assumes you're connecting to the container that holds NHP results.
}
\examples{
\dontrun{azkit::get_container() |> get_nhp_result_sets()}
}
