% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{get_nhp_results}
\alias{get_nhp_results}
\title{Unzip, Read and Parse an NHP Results File}
\usage{
get_nhp_results(
  container_results = Sys.getenv("AZ_STORAGE_CONTAINER_RESULTS"),
  blob_url = Sys.getenv("AZ_STORAGE_EP"),
  file
)
}
\arguments{
\item{container_results}{Name of a blob_container/storage_container object
that stores results files.}

\item{file}{Character. The path to a file in the named \code{container}.}
}
\value{
A nested list.
}
\description{
Unzip, Read and Parse an NHP Results File
}
\details{
Assumes you've connected to the container that holds NHP results.
}
\examples{
\dontrun{
container <- get_container()
result_sets <- container |> get_nhp_result_sets()
file <- result_sets |> dplyr::slice(1) |> dplyr::pull(file)
r <- container |> get_nhp_results(file)
}
}
