% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_scheme_name}
\alias{make_scheme_name}
\title{Concatenate Scheme Name and Code}
\usage{
make_scheme_name(
  scheme_code,
  lookup_path = "supporting_data/scheme-lookup.csv",
  as_filestring = FALSE
)
}
\arguments{
\item{scheme_code}{Character. A focus scheme's three-character ODS code.}

\item{lookup_path}{Character. The file path to the CSV lookup of scheme names
and codes.}

\item{as_filestring}{Logical. Express as a string with punctuation removed,
hyphen-delimited and in lowercase? Used to build filepath.}
}
\value{
Character string.
}
\description{
Concatenate Scheme Name and Code
}
\examples{
\dontrun{construct_scheme_name("XYZ")}
}
