% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_wrap}
\alias{str_wrap}
\title{Wrap long strings to a line length (approximately)}
\usage{
str_wrap(vec, width = 72)
}
\arguments{
\item{vec}{A character vector}

\item{width}{integer The desired line length of the wrapped text (default 72)}
}
\value{
A character vector
}
\description{
A home-made replacement for the core functionality of \code{stringr::str_wrap()}.
}
\details{
Set line length using the \code{width} argument (confusingly) - this matches the
argument name in \code{stringr::str_wrap()} and \code{stringi::stri_wrap()} at least.

This function will find the nearest appropriate breaks to your desired width;
this does however mean that final line lengths may be longer than the
specified width, and if a particularly long word is unfortunately placed, the
resulting line may be much longer than \code{width}.
The default line length is 72 characters, to increase the chance that all
resulting line lengths will be 80 characters at most.
Any existing new line characters in a string will not be retained; they will
be converted to spaces and treated as potential break points. If you wish to
ensure existing new line breaks are kept, replace them with another marker
character before wrapping and then restore them afterwards manually.
}
