% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{get_container}
\alias{get_container}
\title{Connect to an Azure Container}
\usage{
get_container(endpoint, container)
}
\arguments{
\item{tenant}{Character. The tenant ID.}

\item{app_id}{Character. The app ID.}

\item{ep_uri}{Character. The endpoint URI.}

\item{container_name}{Character. The container name. Use \code{Sys.getenv()} with
\code{"AZ_STORAGE_CONTAINER_RESULTS"} or \code{"AZ_STORAGE_CONTAINER_RESULTS"}.}
}
\value{
A blob_container/storage_container object.
}
\description{
Connect to an Azure Container
}
\details{
All arguments default to environmental variables stored in your
.Renviron file. Note that you'll be routed automatically to the browser
for authentication if you don't have a cached token already.
}
\examples{
\dontrun{get_container()}
}
